Revert "Add install_requirements.{sh,py} that only install requirements (#7715)"
This reverts commit cd0e5845cf2c2612b0a6d2a29269043fe3c315f9.
