+++ install_executorch.sh
@@ -7,4 +7,19 @@
 
 # Before doing anything, cd to the directory containing this script.
 cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null || /bin/true
-./run_python_script.sh ./install_executorch.py "$@"
+
+# Find the names of the python tools to use.
+if [[ -z $PYTHON_EXECUTABLE ]];
+then
+  if [[ -z $CONDA_DEFAULT_ENV ]] || [[ $CONDA_DEFAULT_ENV == "base" ]] || [[ ! -x "$(command -v python)" ]];
+  then
+    PYTHON_EXECUTABLE=python3
+  else
+    PYTHON_EXECUTABLE=python
+  fi
+fi
+
+$PYTHON_EXECUTABLE ./install_executorch.py "$@"
+
+# Exit with the same status as the python script.
+exit $?
